back-end/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
fastapi-cors==0.0.6
pydantic-settings==2.1.0
redis==5.0.1
slowapi==0.1.9
email-validator==2.1.0

# Security and monitoring
sentry-sdk[fastapi]==1.38.0
python-json-logger==2.0.7
cryptography==41.0.7
psutil==5.9.6

# Mock data generation dependencies
faker==20.1.0
requests==2.31.0
aiohttp==3.9.1
pyyaml==6.0.1
pillow==10.1.0
minio==7.2.0
